Updated with new functionality
The scoring command has been changed/split. For previous functionality, such as from Part 1, use the "showscore" command, which will print to screen the scores like previously. For future saving of scores, running the "score" command will not print to screen, and will save scores into .txt files in the /output directory. This is needed for uploading, in order to save the individual metric scores